fix(macos): change lastKnownBufferLength to int to fix -Wsign-compare
int comparison against a uint64_t member tripped -Werror,-Wsign-compare
on clang. The value is always a small positive buffer length (assigned
from int, returned as int), so int is the correct type.

+++ src/plugin/plugin.hpp
@@ -148,7 +148,7 @@ namespace ReplayBufferPro
     ReplayBufferManager *replayManager; ///< Replay buffer manager
     HotkeyManager *hotkeyManager;       ///< Hotkey manager
     QTimer *settingsMonitorTimer;       ///< Timer for monitoring OBS settings changes
-    uint64_t lastKnownBufferLength;     ///< Last known buffer length from OBS settings
+    int lastKnownBufferLength;          ///< Last known buffer length from OBS settings
 
     //=========================================================================
     // INITIALIZATION
